refactor: Phase 3 god-object decomposition — extract 5 helper modules
Decompose orchestrator.py (2,673→2,274 LOC, -399) and
optimized_function.py (2,777→1,810 LOC, -967) by extracting focused
helper modules:

- workflow_trace_manager.py: Span collection and backend submission
- cost_estimator.py: Cost estimation and trial cost extraction
- config_state_manager.py: Config lifecycle, persistence, auto-load
- optimization_pipeline.py: Pipeline parameter resolution and evaluator
  construction
- ci_approval.py: CI/CD environment detection and token validation

Also moved _initialize_backend_client to BackendSessionManager and
validate_dataset to orchestrator_helpers.py.

All 1,483 core tests pass. No behavioral changes — pure structural
refactoring following the dependency injection pattern established by
BackendSessionManager.
